# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/employees/{employeeId}/loans": {
      "post": {
        "operationId": "LoanController_create",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoanDto"
              },
              "examples": {
                "default": {
                  "summary": "New loan",
                  "value": {
                    "totalAmount": 12000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Loans"
        ]
      },
      "get": {
        "operationId": "LoanController_findAllForEmployee",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Loans"
        ]
      }
    },
    "/loans": {
      "get": {
        "operationId": "LoanController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "PAID_OFF"
              ]
            }
          },
          {
            "name": "employeeId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Loans"
        ]
      }
    },
    "/loans/{id}": {
      "get": {
        "operationId": "LoanController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Loans"
        ]
      },
      "delete": {
        "operationId": "LoanController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Loans"
        ]
      }
    },
    "/loans/{id}/approve": {
      "patch": {
        "operationId": "LoanController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveLoanDto"
              },
              "examples": {
                "byCount": {
                  "summary": "Split into 6 equal installments",
                  "value": {
                    "numberOfInstallments": 6,
                    "startDate": "2026-08-01"
                  }
                },
                "byAmount": {
                  "summary": "Fixed 2000/month (last absorbs remainder)",
                  "value": {
                    "installmentAmount": 2000,
                    "startDate": "2026-08-01"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Loans"
        ]
      }
    },
    "/payroll/cycles": {
      "post": {
        "operationId": "PayrollController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePayrollCycleDto"
              },
              "examples": {
                "default": {
                  "summary": "July 2026 payroll",
                  "value": {
                    "month": 7,
                    "year": 2026
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payroll"
        ]
      },
      "get": {
        "operationId": "PayrollController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "REVIEW",
                "APPROVED",
                "CLOSED"
              ]
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "example": 7,
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "example": 2026,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payroll"
        ]
      }
    },
    "/payroll/cycles/{id}": {
      "get": {
        "operationId": "PayrollController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payroll"
        ]
      }
    },
    "/payroll/slips/{slipId}": {
      "get": {
        "operationId": "PayrollController_findSlip",
        "parameters": [
          {
            "name": "slipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payroll"
        ]
      }
    },
    "/payroll/cycles/{id}/recalculate": {
      "post": {
        "operationId": "PayrollController_recalculate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payroll"
        ]
      }
    },
    "/payroll/cycles/{id}/review": {
      "patch": {
        "operationId": "PayrollController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payroll"
        ]
      }
    },
    "/payroll/cycles/{id}/approve": {
      "patch": {
        "operationId": "PayrollController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payroll"
        ]
      }
    },
    "/payroll/cycles/{id}/close": {
      "patch": {
        "operationId": "PayrollController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payroll"
        ]
      }
    },
    "/payroll/cycles/{id}/wps": {
      "get": {
        "operationId": "PayrollController_exportWps",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payroll"
        ]
      }
    },
...
    "/reports/dashboard": {
      "get": {
        "operationId": "ReportsController_dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/payroll-summary": {
      "get": {
        "operationId": "ReportsController_payrollSummary",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "example": 7,
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "example": 2026,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/attendance-summary": {
      "get": {
        "operationId": "ReportsController_attendanceSummary",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "example": 7,
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "example": 2026,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/gosi": {
      "get": {
        "operationId": "ReportsController_gosi",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "example": 7,
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "example": 2026,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    }
  },
  "info": {
    "title": "HR System API",
    "description": "Multi-tenant Payroll & HR Management API (Phases 1–6). Import into APIDog via live URL http://localhost:3004/docs-json or backend/postman_collection.json. Swagger UI: http://localhost:3004/docs",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3004/api",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
    "schemas": {
...
      "CreateLoanDto": {
        "type": "object",
        "properties": {
          "totalAmount": {
            "type": "number",
            "example": 12000,
            "description": "Total loan amount (must be positive)."
          }
        },
        "required": [
          "totalAmount"
        ]
      },
      "ApproveLoanDto": {
        "type": "object",
        "properties": {
          "numberOfInstallments": {
            "type": "number",
            "example": 6,
            "description": "Split the total into this many equal monthly installments."
          },
          "installmentAmount": {
            "type": "number",
            "example": 2000,
            "description": "Fixed monthly installment amount (the last one absorbs any remainder)."
          },
          "startDate": {
            "type": "string",
            "example": "2026-08-01",
            "description": "Due date of the first installment; each following one is +1 month."
          }
        },
        "required": [
          "startDate"
        ]
      },
      "CreatePayrollCycleDto": {
        "type": "object",
        "properties": {
          "month": {
            "type": "number",
            "example": 7,
            "minimum": 1,
            "maximum": 12
          },
          "year": {
            "type": "number",
            "example": 2026,
            "minimum": 2000,
            "maximum": 2100
          }
        },
        "required": [
          "month",
          "year"
        ]
      }
    }
  }
}